# spec/openapi.yaml
openapi: "3.1.0"
info:
  title: GPG Chain
  version: "0.1.0"
  description: |
    Distributed append-only ledger of GPG public keys with web-of-trust signing.

    The server stores cryptographic material only. Trust decisions are always client-side.
    Nodes gossip new blocks, signatures, and revocations to peers. The /p2p/* endpoints
    are used exclusively for inter-node communication.

    All errors are returned as {"error": "..."} with an appropriate HTTP status code.
    Fingerprints are uppercase hex with no spaces. All signatures are base64-encoded
    detached binary OpenPGP signatures. Signing payloads are typed binary strings as
    defined in spec/payloads.md.
paths:
...
  /block/{fingerprint}:
    get:
      summary: Get a single block by fingerprint
      description: |
        Returns the full block for the given fingerprint, including its complete sig chain
        (all SigEntry records in order from oldest to newest) and revocation status.
      parameters:
        - name: fingerprint
          in: path
          required: true
          description: Uppercase hex fingerprint of the key (40 chars for v4, 64 for v5).
          schema:
            type: string
            pattern: "^[A-F0-9]{40}$|^[A-F0-9]{64}$"
      responses:
        "200":
          description: The block and its full sig chain.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "404":
          description: No block with that fingerprint exists on this node.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /block:
    post:
      summary: Submit a new key block
      description: |
        Adds a new GPG public key to the ledger. The submitter must provide a self-signature
        proving they control the corresponding private key.

        The node:
        1. Parses the armored_key and extracts the fingerprint and UIDs.
        2. Enforces minimum key strength (RSA >= 2048 bit; Ed25519 accepted; DSA-1024 rejected).
        3. Checks that at least one UID email domain matches the node's configured allowlist.
        4. Extracts the submit_timestamp from the self_sig by verifying the SUBMIT payload.
        5. Verifies the self_sig against the GPGCHAIN_SUBMIT_V1 payload.
        6. Computes and stores the block hash.
        7. Gossips the new block to K random peers.

        The node does not accept a timestamp field from the client request; the timestamp
        is extracted from the signed SUBMIT payload, which commits to it. The node stores
        whatever timestamp the client embedded in the signed payload.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBlockRequest"
      responses:
        "201":
          description: Block created and stored. Returns the full block as stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "400":
          description: |
            Invalid request. Possible reasons: malformed armored key, signature verification
            failure, key strength below minimum, no UID found, invalid base64 in self_sig.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: |
            The key's UID email domains do not match this node's configured domain allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A block with this fingerprint already exists on this node.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /block/{fingerprint}/sign:
    post:
      summary: Append a trust signature to a block
      description: |
        Appends a new SigEntry to the sig chain of the identified block. The signer may be
        on-ledger (their block exists on this node) or off-ledger (they provide their armored
        key inline).

        The node:
        1. Looks up the target block by fingerprint. Returns 404 if not found.
        2. Returns 409 if the block is revoked.
        3. Looks up the signer's key: from the local ledger if on-ledger, or from
           signer_armored_key if off-ledger.
        4. Enforces minimum key strength on the signer's key.
        5. Verifies the sig against the GPGCHAIN_TRUST_V1 payload.
        6. Returns 409 if this signer_fingerprint has already signed this block (duplicate signer).
        7. Constructs and appends the SigEntry. Sets prev_hash to the current sig_chain_head
           (or to the block's hash if this is the first signature). Computes and stores the
           SigEntry hash.
        8. Updates sig_chain_head on the block.
        9. Gossips the new SigEntry to K random peers.
      parameters:
        - name: fingerprint
          in: path
          required: true
          description: Fingerprint of the block being signed.
          schema:
            type: string
            pattern: "^[A-F0-9]{40}$|^[A-F0-9]{64}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignBlockRequest"
      responses:
        "200":
          description: SigEntry appended successfully. Returns the updated block.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "400":
          description: |
            Invalid request. Possible reasons: signature verification failure, signer key
            too weak, malformed signer_armored_key, signer_fingerprint does not match
            signer_armored_key, invalid base64 in sig.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No block with that fingerprint exists on this node.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: |
            Either: this signer_fingerprint has already signed this block (duplicate signer),
            or the target block is revoked and cannot receive new signatures.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /block/{fingerprint}/revoke:
    post:
      summary: Revoke a block (owner only)
      description: |
        Permanently revokes a block. Only the key owner (the entity controlling the private
        key in the block's armored_key) can revoke.

        The node:
        1. Looks up the block by fingerprint. Returns 404 if not found.
        2. Returns 409 if already revoked.
        3. Verifies the sig against the GPGCHAIN_REVOKE_V1 payload using the block's own key.
        4. Returns 403 if verification fails (signature is not from the block's own key).
        5. Sets revoked=true, stores revocation_sig, gossips the revocation to K peers.

        Revocation is permanent and irreversible. A revoked block remains in the ledger
        with revoked=true and will never accept new trust signatures.
      parameters:
        - name: fingerprint
          in: path
          required: true
          description: Fingerprint of the block to revoke.
          schema:
            type: string
            pattern: "^[A-F0-9]{40}$|^[A-F0-9]{64}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeBlockRequest"
      responses:
        "200":
          description: Block revoked successfully. Returns the updated block.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "400":
          description: |
            Invalid request. Possible reasons: malformed sig, invalid base64.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: |
            The revocation signature does not verify against the block's own key. The caller
            is not the key owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No block with that fingerprint exists on this node.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The block is already revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /search:
    get:
      summary: Search blocks by UID substring
      description: |
        Returns all blocks whose UIDs contain the query string as a case-insensitive substring.
        The search is performed across all UID strings for each block.
        Always returns HTTP 200. Returns an empty array if no blocks match.
        Revoked blocks are included in results (clients filter based on revoked field if desired).
      parameters:
        - name: q
          in: query
          required: true
          description: Substring to search for within block UIDs (case-insensitive).
          schema:
            type: string
      responses:
        "200":
          description: Array of matching blocks (may be empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Block"
...
  /peers:
    get:
      summary: List known peer URLs
      description: |
        Returns the list of peer URLs this node currently knows about. Includes all peers
        regardless of health status.
      responses:
        "200":
          description: Array of peer URL strings.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uri
                  example: "https://keys.partner.org"

    post:
      summary: Register a new peer
      description: |
        Registers a new peer URL with this node. The node applies acceptance rules:
        1. URL scheme must be http or https.
        2. Host must not resolve to a private or loopback address.
        3. Reciprocal reachability check: GET /peers on the claimed URL must return 200.
        4. Peer list must not be at capacity (default max 50).
        5. Duplicate URLs are silently accepted (return 200, not added twice).

        On success, the receiving node also attempts to register itself with the new peer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPeerRequest"
      responses:
        "200":
          description: Peer registered (or already present).
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        "400":
          description: |
            Invalid URL, disallowed URL scheme, private/loopback address, or reciprocal
            reachability check failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Peer list is at capacity. No new peers can be added.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /p2p/block/{hash}:
    get:
      summary: Fetch a single block by its content hash
      description: |
        Returns the full block identified by its content hash (not its fingerprint). Used
        by peers during sync to fetch a specific block they know the hash of.
        Returns the block with its complete sig chain.
      parameters:
        - name: hash
          in: path
          required: true
          description: The content hash of the block (64-char uppercase hex SHA-256).
          schema:
            type: string
            pattern: "^[A-F0-9]{64}$"
      responses:
        "200":
          description: The full block.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "404":
          description: No block with that hash exists on this node.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /p2p/block:
    post:
      summary: Receive a gossiped block from a peer
      description: |
        Accepts a block gossiped by a peer. Applies the same validation as POST /block:
        self-sig verification, hash verification, key strength, domain allowlist.
        If valid and not already seen, stores the block and gossips it to K more peers.
        If already seen, returns 200 without re-storing or re-gossiping.
        If invalid, returns 400 and does not store or forward.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GossipBlockRequest"
      responses:
        "200":
          description: Block accepted (or already known). Stored and gossiped if new.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        "400":
          description: |
            Invalid block. Validation failed: bad signature, hash mismatch, weak key,
            domain not allowed, or malformed data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /p2p/sign:
    post:
      summary: Receive a gossiped trust signature from a peer
      description: |
        Accepts a SigEntry gossiped by a peer. Applies full validation: SigEntry hash
        verification, signature verification against TRUST payload, prev_hash link check.
        The target block must already exist on this node.
        If valid and not already seen, stores the SigEntry and gossips it to K more peers.
        If already seen, returns 200 without re-storing or re-gossiping.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GossipSignRequest"
      responses:
        "200":
          description: SigEntry accepted (or already known).
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        "400":
          description: |
            Invalid SigEntry: hash mismatch, signature verification failure, missing target
            block, malformed data, or duplicate signer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    AddBlockRequest:
      type: object
      description: Request body for POST /block. The node extracts the fingerprint from the key.
      required:
        - armored_key
        - self_sig
      properties:
        armored_key:
          type: string
          description: |
            Full ASCII-armored OpenPGP public key block to register on the ledger.
          example: "-----BEGIN PGP PUBLIC KEY BLOCK-----\n...\n-----END PGP PUBLIC KEY BLOCK-----\n"
        self_sig:
          type: string
          description: |
            Base64-encoded detached binary OpenPGP signature over the GPGCHAIN_SUBMIT_V1
            payload. Must be signed by the private key corresponding to armored_key.
          example: "iQEzBAABCAAdFiEEo/nC4bjUdlDynjwee4TWb1CeKBwFAmYJH..."
...
    RevokeBlockRequest:
      type: object
      description: Request body for POST /block/{fingerprint}/revoke.
      required:
        - sig
      properties:
        sig:
          type: string
          description: |
            Base64-encoded detached binary OpenPGP signature over the GPGCHAIN_REVOKE_V1
            payload. Must be signed by the private key corresponding to the block being
            revoked (i.e. the key in the block's armored_key field).
          example: "iQEzBAABCAAdFiEEo/nC4bjUdlDynjwee4TWb1CeKBwFAmYJR..."
    AddPeerRequest:
      type: object
      description: Request body for POST /peers.
      required:
        - addr
      properties:
        addr:
          type: string
          format: uri
          description: |
            URL of the peer node to register. Must use http or https scheme. Must not
            resolve to a private or loopback address.
          example: "https://keys.partner.org"
...
    ErrorResponse:
      type: object
      description: Returned by all endpoints on error.
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable description of the error.
          example: "signature verification failed"
...